Refactor alpine models (#72)
* refactor: audio-player bundle

* refactor: bundle base and faqs view models

* refactor: bundle medium view model

* refactor: devotion banner bundle

* refactor: bundle error page model

* refactor: bundle testimonials slider

* refactor: bundle mailer lite popup

* chore: clean-up data file
