+++ .gitignore
@@ -2,3 +2,4 @@ build
 node_modules
 .cache
 mix-manifest.json
+.vscode
