+++ src/_blocks/channels.njk
@@ -1,6 +1,6 @@
 <section
   class="lg:container mx-auto p-4"
-  x-data="app.mediumViewModel"
+  x-data="mediumView"
   id="start"
 >
   <div class="relative lg:rounded">
