+++ src/_data/channels.json
@@ -13,7 +13,7 @@
         "description": "Registriere dich jetzt, um \"Bibel in Einem Jahr\" jeden Morgen in deinem Posteingang zu haben. Du bekommst jeden Tag eine E-Mail.",
         "image": "/assets/img/medium_email.jpg",
         "cta": {
-          "code": "<a href=\"https://landing.mailerlite.com/webforms/landing/b4j7s5\" class=\"inline-block bg-white rounded p-2 pb-1.5 text-red hover:bg-red hover:text-white transition-colors js-mailerLite \"data-ml=\"5166542,b4j7s5\">Jetzt registrieren</a>"
+          "code": "<a @click=\"mailerLitePopup\" href=\"https://landing.mailerlite.com/webforms/landing/b4j7s5\" class=\"inline-block bg-white rounded p-2 pb-1.5 text-red hover:bg-red hover:text-white transition-colors  \"data-ml=\"5166542,b4j7s5\">Jetzt registrieren</a>"
         }
       },
       {
@@ -47,7 +47,7 @@
         "description": "Sign up now to receive Bible in One Year in your inbox each morning. You’ll get one email each day.",
         "image": "/assets/img/medium_email.jpg",
         "cta": {
-          "code": "<a href=\"https://landing.mailerlite.com/webforms/landing/y7w6m3\" class=\"inline-block bg-white rounded p-2 pb-1.5 text-red hover:bg-red hover:text-white transition-colors js-mailerLite \"data-ml=\"5166494,y7w6m3\">अभी साइनअप करें</a>"
+          "code": "<a @click=\"mailerLitePopup\" href=\"https://landing.mailerlite.com/webforms/landing/y7w6m3\" class=\"inline-block bg-white rounded p-2 pb-1.5 text-red hover:bg-red hover:text-white transition-colors  \"data-ml=\"5166494,y7w6m3\">अभी साइनअप करें</a>"
         }
       },
       {
@@ -77,7 +77,7 @@
         "description": "Sign up now to receive Bible in One Year in your inbox each morning. You’ll get one email each day.",
         "image": "/assets/img/medium_email.jpg",
         "cta": {
-          "code": "<a href=\"https://landing.mailerlite.com/webforms/landing/p5d3p8\" class=\"inline-block bg-white rounded p-2 pb-1.5 text-red hover:bg-red hover:text-white transition-colors js-mailerLite \"data-ml=\"4010371,p5d3p8\">Sign up now</a>"
+          "code": "<a @click=\"mailerLitePopup\" href=\"https://landing.mailerlite.com/webforms/landing/p5d3p8\" class=\"inline-block bg-white rounded p-2 pb-1.5 text-red hover:bg-red hover:text-white transition-colors  \"data-ml=\"4010371,p5d3p8\">Sign up now</a>"
         }
       },
       {
@@ -117,7 +117,7 @@
         "description": "Inscrivez-vous maintenant pour recevoir la Bible en 1 an dans votre boîte de réception chaque matin. Vous recevrez un e-mail par jour.",
         "image": "/assets/img/medium_email.jpg",
         "cta": {
-          "code": "<a href=\"https://landing.mailerlite.com/webforms/landing/d4q3b1\" class=\"inline-block bg-white rounded p-2 pb-1.5 text-red hover:bg-red hover:text-white transition-colors js-mailerLite \"data-ml=\"5166560,d4q3b1\">S'inscrire maintenant</a>"
+          "code": "<a @click=\"mailerLitePopup\" href=\"https://landing.mailerlite.com/webforms/landing/d4q3b1\" class=\"inline-block bg-white rounded p-2 pb-1.5 text-red hover:bg-red hover:text-white transition-colors  \"data-ml=\"5166560,d4q3b1\">S'inscrire maintenant</a>"
         }
       },
       {
@@ -145,7 +145,7 @@
         "description": "Receive a daily email with a fresh devotion straight in your inbox.",
         "image": "/assets/img/medium_app.jpg",
         "cta": {
-          "code": "<a href=\"https://landing.mailerlite.com/webforms/landing/p9r9p0\" class=\"inline-block bg-white rounded p-2 pb-1.5 text-red hover:bg-red hover:text-white transition-colors js-mailerLite\" data-ml=\"4091911,p9r9p0\">Regístrate ahora</a>"
+          "code": "<a @click=\"mailerLitePopup\" href=\"https://landing.mailerlite.com/webforms/landing/p9r9p0\" class=\"inline-block bg-white rounded p-2 pb-1.5 text-red hover:bg-red hover:text-white transition-colors \" data-ml=\"4091911,p9r9p0\">Regístrate ahora</a>"
         }
       },
       {
@@ -176,7 +176,7 @@
         "description": "Sign up now to receive Bible in One Year in your inbox each morning. You’ll get one email each day.",
         "image": "/assets/img/medium_email.jpg",
         "cta": {
-          "code": "<a href=\"https://landing.mailerlite.com/webforms/landing/q8q8x7\" class=\"inline-block bg-white rounded p-2 pb-1.5 text-red hover:bg-red hover:text-white transition-colors js-mailerLite \"data-ml=\"5166785,q8q8x7\">立即注册</a>"
+          "code": "<a @click=\"mailerLitePopup\" href=\"https://landing.mailerlite.com/webforms/landing/q8q8x7\" class=\"inline-block bg-white rounded p-2 pb-1.5 text-red hover:bg-red hover:text-white transition-colors  \"data-ml=\"5166785,q8q8x7\">立即注册</a>"
         }
       },
       {
@@ -200,7 +200,7 @@
         "description": "سجل الآن لتحصل على الكتاب المقدس في عام واحد على بريدك الإلكتروني كل صباح. ستتلقى بريدًا إلكترونيًا واحدًا كل يوم.",
         "image": "/assets/img/medium_email.jpg",
         "cta": {
-          "code": "<a href=\"https://landing.mailerlite.com/webforms/landing/l2u8i1\" class=\"inline-block bg-white rounded p-2 pb-1.5 text-red hover:bg-red hover:text-white transition-colors js-mailerLite \"data-ml=\"5166407,l2u8i1\">سجل اليوم</a>"
+          "code": "<a @click=\"mailerLitePopup\" href=\"https://landing.mailerlite.com/webforms/landing/l2u8i1\" class=\"inline-block bg-white rounded p-2 pb-1.5 text-red hover:bg-red hover:text-white transition-colors  \"data-ml=\"5166407,l2u8i1\">سجل اليوم</a>"
         }
       },
       {
@@ -234,7 +234,7 @@
         "description": "Sign up now to receive Bible in One Year in your inbox each morning. You’ll get one email each day.",
         "image": "/assets/img/medium_email.jpg",
         "cta": {
-          "code": "<a href=\"https://landing.mailerlite.com/webforms/landing/a0u5h1\" class=\"inline-block bg-white rounded p-2 pb-1.5 text-red hover:bg-red hover:text-white transition-colors js-mailerLite \"data-ml=\"5166713,a0u5h1\">สมัครตอนนี้</a>"
+          "code": "<a @click=\"mailerLitePopup\" href=\"https://landing.mailerlite.com/webforms/landing/a0u5h1\" class=\"inline-block bg-white rounded p-2 pb-1.5 text-red hover:bg-red hover:text-white transition-colors  \"data-ml=\"5166713,a0u5h1\">สมัครตอนนี้</a>"
         }
       },
       {
@@ -259,7 +259,7 @@
         "description": "Sign up now to receive Bible in One Year in your inbox each morning. You’ll get one email each day.",
         "image": "/assets/img/medium_email.jpg",
         "cta": {
-          "code": "<a href=\"https://landing.mailerlite.com/webforms/landing/a9d3s5\" class=\"inline-block bg-white rounded p-2 pb-1.5 text-red hover:bg-red hover:text-white transition-colors js-mailerLite \"data-ml=\"5166776,a9d3s5\">Sign up now</a>"
+          "code": "<a @click=\"mailerLitePopup\" href=\"https://landing.mailerlite.com/webforms/landing/a9d3s5\" class=\"inline-block bg-white rounded p-2 pb-1.5 text-red hover:bg-red hover:text-white transition-colors  \"data-ml=\"5166776,a9d3s5\">Sign up now</a>"
         }
       },
       {
