+++ src/_includes/layouts/base.njk
@@ -43,7 +43,7 @@
   </head>
 
   <body
-    x-data="app.baseViewModel"
+    x-data="baseView"
     @keydown.escape="isLangPickerOpen = false"
   >
