+++ src/_includes/partials/audio-player.njk
@@ -5,7 +5,7 @@
     {% endif %}
 {% endfor %}
 
-<section x-data="app.audioPlayerModel" x-init="getChapters({{ chapters | dump}})" >
+<section x-data="audioPlayer" x-init="getChapters({{ chapters | dump}})" >
     {# Circlular Progress Bar: https://css-tricks.com/building-progress-ring-quickly/ #}
     <button x-cloak x-show="!isPlaying" class="absolute ltr:right-1 rtl:left-0 bottom-4 inline-flex items-center justify-center overflow-hidden rounded-full hover:bg-red/10 active:bg-red/20 focus:outline-none focus-visible:ring focus:ring-red/10 transition duration-100 ease-out hover:ease-in active:ease-in bg-white shadow-md" @click="showAudioPlayer()">
         <svg class="h-20 w-20 -rotate-90 rounded-full">
