+++ src/assets/js/mediumViewModel.js
@@ -59,5 +59,35 @@ export default () => {
         previousPodcast.addEventListener("click", () => podcastSwiper.prev());
       }
     },
+    mailerLitePopup(e) {
+      (function (m, a, i, l, e, r) {
+        m["MailerLiteObject"] = e;
+        function f() {
+          var c = { a: arguments, q: [] };
+          var r = this.push(c);
+          return "number" != typeof r ? r : f.bind(c.q);
+        }
+        f.q = f.q || [];
+        m[e] = m[e] || f.bind(f.q);
+        m[e].q = m[e].q || f.q;
+        r = a.createElement(i);
+        var _ = a.getElementsByTagName(i)[0];
+        r.async = 1;
+        r.src = l + "?v" + ~~(new Date().getTime() / 1000000);
+        _.parentNode.insertBefore(r, _);
+      })(
+        window,
+        document,
+        "script",
+        "https://static.mailerlite.com/js/universal.js",
+        "ml"
+      );
+
+      let ml_account = ml("accounts", "3056200", "f0i3r5m4t5", "load");
+
+      e.preventDefault();
+      let d = e.target.dataset.ml.split(",");
+      ml_account("webforms", String(d[0]), String(d[1]), "show");
+    },
   };
 };
