+++ src/faqs/faqs.njk
@@ -23,7 +23,7 @@ redirectTo: faqs
 
 <div
   class="mx-auto py-2 sm:py-4 lg:container"
-  x-data="app.faqsViewModel"
+  x-data="faqsView"
 >
   <div class="js-filter hidden flex relative mb-4">
     <span class="py-2 pr-2 font-bold absolute ltr:left-0 rtl:right-0 pointer"
